+++ docs-source/changelog.md
@@ -4,6 +4,12 @@ v1.800 (still open)
 
 - [BREAKING] Rotations are now stored and managed as transformations. Euler
   angles (x, y, z rotation) support remains but with a modified interface.
+- [BREAKING] SceneObject.offset has been renamed to SceneObject.position.
+- [BREAKING] SceneObject.boundingBoxSet is now private. It is still possible
+  to retrieve the bounding boxes as Models, using SceneObject.getBoundingBoxSetModels.
+- SceneObject.boundingBoxSet and SceneObject.models have been converted
+  to smart pointers to allow for the copy constructor to create scene
+  objects that reuse existing models.
 
 v1.709 2021-10-16